Lab2 3.1.1 New class IntArrayStack
